+++ src/lib.rs
@@ -5,13 +5,12 @@
 //! Lightweight diagnostic utilities for embedded Rust.
 //! Compatible with `no_std` and multiple HAL backends.
 
-pub mod logger;
 pub mod scanner;
 
 #[macro_use]
 pub mod macros;
 
 pub mod compat;
+pub mod explore;
 pub mod error;
-pub mod explorer;
 pub mod prelude;
